refactor: Split ContactCardModal in different components
nothing changed, just refactoring
